move tests to AgentTest.java
